+++ src/main/kotlin/com/padovese/assessment/exception/NegativeArgException.kt
@@ -1,6 +1,6 @@
 package com.padovese.assessment.exception
 
 
-class NegativeArgException(arg: String, value: Long) : RuntimeException() {
+class NegativeArgException(arg: String, value: Int) : RuntimeException() {
     override val message: String? = "The requesting $arg is $value. It must not be negative."
 }
